making MemoryAddress an Unsigned Int

+++ compiler/test/TestMemory.kt
@@ -387,12 +387,12 @@ class TestMemory: FunSpec({
             target.memorySize(BaseDataType.BYTE) shouldBe 1
             target.memorySize(BaseDataType.WORD) shouldBe 2
             target.memorySize(BaseDataType.LONG) shouldBe 4
-            target.memorySize(BaseDataType.FLOAT) shouldBe target.FLOAT_MEM_SIZE
+            target.memorySize(BaseDataType.FLOAT) shouldBe target.FLOAT_MEM_SIZE.toInt()
 
             target.memorySize(DataType.BOOL, null) shouldBe 1
             target.memorySize(DataType.WORD, null) shouldBe 2
             target.memorySize(DataType.LONG, null) shouldBe 4
-            target.memorySize(DataType.FLOAT, null) shouldBe target.FLOAT_MEM_SIZE
+            target.memorySize(DataType.FLOAT, null) shouldBe target.FLOAT_MEM_SIZE.toInt()
 
             target.memorySize(DataType.STR, null) shouldBe 2
             target.memorySize(DataType.STR, 50) shouldBe 50
@@ -405,14 +405,14 @@ class TestMemory: FunSpec({
             target.memorySize(DataType.arrayFor(BaseDataType.WORD), 10) shouldBe 20
             target.memorySize(DataType.arrayFor(BaseDataType.UWORD), 10) shouldBe 20
             target.memorySize(DataType.arrayFor(BaseDataType.LONG), 10) shouldBe 40
-            target.memorySize(DataType.arrayFor(BaseDataType.FLOAT), 10) shouldBe 10*target.FLOAT_MEM_SIZE
+            target.memorySize(DataType.arrayFor(BaseDataType.FLOAT), 10) shouldBe 10*target.FLOAT_MEM_SIZE.toInt()
             target.memorySize(DataType.arrayFor(BaseDataType.WORD, true), 10) shouldBe 20
             target.memorySize(DataType.arrayFor(BaseDataType.UWORD, true), 10) shouldBe 20
 
             target.memorySize(DataType.BOOL, 10) shouldBe 10
             target.memorySize(DataType.UWORD, 10) shouldBe 20
             target.memorySize(DataType.LONG, 10) shouldBe 40
-            target.memorySize(DataType.FLOAT, 10) shouldBe 10*target.FLOAT_MEM_SIZE
+            target.memorySize(DataType.FLOAT, 10) shouldBe 10*target.FLOAT_MEM_SIZE.toInt()
         }
     }
